requirements.txt
Flask==0.12.2
requests==2.13.0
python-oauth2==1.0.1
